sw: Backspace/Delete hidden section must delete flys anchored therein
During save to ODF this asserts, or later crashes:
    atrfrm.cxx:3810: void sw::CheckAnchoredFlyConsistency(const SwDoc&): Assertion `std::find(rFlys.begin(), rFlys.end(), pSpz) != rFlys.end()' failed.

Initial idea was to fix this by changing SwWrtShell to not use the
ArtificialSelection mode when a hidden section is selected - this fixes
the initial problem but there is another one!

Because ArtificialSelection is now also set in
SwXTextRange::DeleteAndInsert() and SwXTextCursor::DeleteAndInsert(),
the crash can also be reproduced by calling setString("nonempty") on a
cursor that selects multiple nodes with an object anchored on a middle
node.

So better to fix this in SwUndoSaveContent::DelContentIndex() by moving
the Replace check into the IsSelected functions, and to ignore *only*
flys on the start and the end node of the selection, so flys anchored on
the middle nodes are now deleted.

In any case for Undo to work a SwHistoryHint must be created for these
which ExcludeFlyAtStartEnd prevents so it's a different case.

This covers all the initially intended use-cases of
DelContentType::Replace, which actually only were 1-2 node selections
anyway (limited by UI); the arbitrary-sized selection cases were from
later re-uses of the flag for other use cases, and in both the Backspace
and the API case the possibility of flys anchored on middle nodes was
unfortunately missed.

Currently SwUndoSaveContent::DelContentIndex() was the only caller of
the IsSelected functions that could see a DelContentType::Replace.

(regression from commit 85376a02348810812d515ee72140dbf56f2b6040)

Change-Id: Ifc5aad47e80c638d4e7cb5f66c035fdf1e34b89f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188251
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>
(cherry picked from commit 40d659647d3dbb5e316963968947272f0a71104d)