vcl dnd: Consolidate drop target retrieval, avoid casting
Return a reference to the concrete DNDListenerContainer
instance instead of one to the abstract
css::datatransfer::dnd::XDropTarget interface in
Window::GetDropTarget.

This allows to drop Window::GetDragGestureRecognizer
which was returning a reference to the
css::datatransfer::dnd::XDragGestureRecognizer interface
for the exact same object, so can now use
Window::GetDropTarget directly instead, which drops the
need to query for the other interface.

DNDEventDispatcher was already assuming that a
DNDListenerContainer object was returned and
casting to that, which is no longer needed now.

Change-Id: Ib7ea7cdb0650ba2410b112774f475b4799901538
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188271
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins