tdf#167661: collect bullet fonts before exporting font table
The problem was, that InitFontTable, used to collect all fonts used
in the document, didn't handle bullets. When font table was written,
some fonts needed for bullets could be still missing from it; only
when writing list levels, these fonts were added to maFonts.

It might seem that that is a bug in SwDoc::ForEachCharacterFontItem,
called from InitFontTable. It might be; and maybe even needs to be
addressed at some point (though I don't know which side effects it
could have). But in the case of export to RTF, this wouldn't fix it
completely, because there is some special processing of bullets in
MSWordExportBase::SubstituteBullet, which might change the exported
font name; therefore, it could happen to still be not covered by an
iteration over actual document's fonts.

1. Extract the code filling used numbering rules in GetNumberingId,
into a separate function (EnsureUsedNumberingTable).

2. Extract the code getting bullet string + font in NumberingLevel,
into a separate function (GetNumberingLevelBulletStringAndFont).

3. In wwFontHelper::InitFontTable, use these functions to iterate
the used numbering rules, and handle fonts of these. To do that,
let the function take an Export object.

Change-Id: Ib7ecdeb9908cd3bd53346ccb035d623fa603db0d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188277
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>