sc: add a way to remove the current sheet view
If we are currently in a sheet view, this command will remove the
sheet view (including the ScTable representing the sheet view),
and return to the default view.
Also add state code to not allow adding or removing a sheet view
when read-only and disable remove command when in default view.

Change-Id: I1b44db9faa717d53631b1e0571d4dba1e9856a5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188467
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>