tdf#130857 qt weld: Support setting TreeView "expander image"
Support the special index of -1 in the various
QtInstanceTreeView::set_image variants, which
is documented in weld.hxx as

    // col index -1 sets the expander image

GtkTreeViewColumn supports multiple renderers
for one column, e.g. to set an image and text
in addition, which is used e.g. in Writer's
"Manage Changes" dialog
(UI file: svx/uiconfig/ui/redlineviewpage.ui,
insertion of rows in SwRedlineAcceptDlg::InsertChildren)
for the first column to hold an "expander image", accessed
by index -1, and text.

QAbstractItemModel (and subclasses) also support containing
multiple types of data for one column, using the concept
of roles, see the QAbstractItemModel::setData doc [1].

For now, simply set the "expander image" as the image
(data for role Qt::DecorationRole) for the first column.

This will be fine as long as only non-image data (like
text) will be set for that column otherwise. If there's
any case where multiple images (i.e. GtkCellRendererPixBuf
in the .ui file) are used inside a single column
(GtkTreeViewColumn in the .ui file), setting either will
overwrite the other.
If that turns out to be a problem in practice, it will
require another solution then.

Avoiding to use multiple images in one column by splitting that
column into multiple ones might be one approach, see e.g.

    commit 8e7aa89201c7156104ee95f5a3c58e3e44a65be3
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jul 21 18:56:15 2025 +0200

        tdf#130857 Use explicit column for image in formula dlg "Struct" page

, but that can't simply be done for the case of the "Manage
Changes" dialog mentioned above, because it has a column
title set that should be set for a single column containing
multiple types of data, not two columns.

[1] https://doc.qt.io/qt-6/qabstractitemmodel.html#setData

Change-Id: Ia4141a221b1da0c475727791da36664f49e31678
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188312
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins