tdf#167669 gtk: Use native GtkColorChooserDialog
Following

    Change-Id: I29380607cdf3abe739a245aeb25526f643dfa537
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jul 25 14:26:06 2025 +0200

        tdf#167669 weld: Introduce weld::ColorChooserDialog

, switch the GTK implementation to use a native
GtkColorChooserDialog [1] for the weld::ColorChooserDialog
implementation instead of using LibreOffice's custom
GtkColorChooserDialog.

Set GtkColorChooserDialog::show-editor [2] to true in
order to not show a set of preselected colors on start,
but show the more advanced control (that can otherwise be
opened by clicking the "+" button) right away.
This one is more similar to the LO one, and e.g. the
"Font Color" popup in the Writer Formatting toolbar already
provides a set of predefined colors to select from, so
not showing another set when clicking the "Custom Color"
button seems reasonable.

Besides providing a more uniform experience across applications,
using the native GtkColorChooserDialog also provides a color
picker to select a color from a pixel on screen, which the
custom LO implementation doesn't have.

In the ColorDialog ctor, explicitly set the dialog to modal.
This is already the case for the custom dialog used so far
(and still used by anything except gtk3/gtk4 after this commit)
as defined in the .ui file (cui/uiconfig/ui/colorpickerdialog.ui),
but it's not the case for GtkColorChooserDialog by default.

[1] https://docs.gtk.org/gtk3/class.ColorChooserDialog.html
[2] https://docs.gtk.org/gtk3/property.ColorChooserDialog.show-editor.html

Change-Id: If5bfb453b55c905ffe332707e2fb261a48664c31
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188344
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins