tdf#167669 tdf#130857 qt: Use native QColorDialog
Similar to how previous commit

    Change-Id: If5bfb453b55c905ffe332707e2fb261a48664c31
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jul 25 14:55:34 2025 +0200

        tdf#167669 gtk: Use native GtkColorChooserDialog

introduced the use of the native GtkColorChooserDialog
for gtk3 and gtk4, use the native QColorDialog [1]
for the Qt-based VCL plugins.

Like GtkColorChooserDialog, it also has the advantage
of providing a color picker to select a color from a pixel
on screen, which the custom LO implementation doesn't have.

[1] https://doc.qt.io/qt-6/qcolordialog.html

Change-Id: If371d75c8a2687430a89de798d50aa10b8db1c29
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188345
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>