tdf#130857 a11y: Set mnemonic widgets in "Theme Color Edit" dialog
... instead of manually setting the label-for and labelled-by
a11y relations.

This ensure that moving focus to one of the
menu buttons is possible using the accelerator key
for the gtk3 VCL plugin, and also that the a11y relations
are set for the qt6 VCL plugin on Linux when using native Qt
widgets with SAL_VCL_QT_USE_WELDED_WIDGETS=1.

To trigger the dialog from Writer:
"Format" -> "Theme" -> "Add"

Change-Id: I15d7279b42c8c3bc54b4a771ca05553ef3178561
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188375
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins