tdf#130857 Define widgets in "Theme Color Edit" dialog in order
Switch the order in which the widgets in the .ui file
are defined so that the order matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

To trigger the dialog from Writer:
"Format" -> "Theme" -> "Add"

Change-Id: I7a528c403b9683b6ad21eb85c4819a5775fc48ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188376
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins