svx: Set proper parent for "Theme Color Edit" dialog
The parent for this modal dialog is the "Theme" dialog
from which the "Theme Color Edit" dialog is started,
and not the "Theme" dialog's parent.

Setting the proper parent e.g. ensures that dialog
modality on top of the parent works as expected with
qt6 and gtk3, while it was previously not enforced
that the dialog was always shown on top of the "Theme"
dialog.

Change-Id: Id9655cf84d6b53edcae9b1575f34f9cfbcd33dd8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188377
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins