tdf#130857 weld: Implement Dialog::set_default_response in base class
There is no need for each weld::Dialog implementation
to implement this separately, because it can be implemented
using weld::Dialog::weld_button_for_response and
weld::Dialog::change_default_button, as done for
QtInstanceButton in previous commit

    Change-Id: I4da34ee14436eb2025781947309cecb397ace4c5
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Jul 26 08:06:56 2025 +0200

        tdf#130857 qt weld: Implement QtInstanceDialog::set_default_response

Therefore, move that implementation to the base class
and use it for both, the Qt and GTK implementations.
(This implies getting rid of one use of GtkDialog API
deprecated in GTK 4.10, [1]).

In fact, GTK 4 also implements gtk_dialog_set_default_response
using gtk_window_set_default_widget internally, see [2].

Leave the SalInstanceDialog::set_default_response override
because the logic in there (and the vcl Dialog class) is
more special and the base class implementation would not
be equivalent, as seen in failing UI test tdf115853 with
an earlier version of this change [3].

[1] https://docs.gtk.org/gtk4/method.Dialog.set_default_response.html
[2] https://gitlab.gnome.org/GNOME/gtk/-/blob/1af8d7bc04852306a5d68a46cfc3ec365790db4d/gtk/deprecated/gtkdialog.c#L1028-1056
[3] https://ci.libreoffice.org/job/gerrit_linux_clang_dbgutil/186869/console

Change-Id: Icacd19b465b88c0d6601b7400a4b5aa59975a6cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188387
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins