tdf#167569: Reimplement RTF character formatting output
The rules for outputting character-level formatting to RTF are utterly
convoluted. The formatting may be output in different context (like in
style definitions, runs, etc.);  and it consists of two distinct parts:
some keywords are independent of current run's script; other depend on
the script, and may require either a "normal" keyword, like \fN, or an
"associated" keyword, such as \afN.  The order of script group keyword
blocks (\loch, \dbch, etc.) is also important.  It is documented (very
poorly) in "Rich Text Format (RTF) Specification, Version 1.9.1", sect.
Font (Character) Formatting Properties.

Previously, we collected character formatting, encoded as RTF keywords,
in the order of iteration of text properties in the export code.

Commit f97af19460fbd7483a0e1c1d0137e814f5390e69
       tdf#137894 separate associated character properties
       2020-11-25
had split keywords of different script groups into separate strings in
an attempt to combine them eventually in a sane order according to the
script. However, the buffers were still filled by the keywords as they
appeared in iteration, when the script of the run was still unknown. A
decision was to always emit \loch keywords as "normal", and all others
as "associated" - which was wrong e.g. for RTL and CJK runs.  It could
happen to look better in LibreOffice in some cases; but it would still
look differently in Word. This can be seen in the bugdoc in tdf#141671,
which looked OK in LO after the fix to that bug, but not in Word.

This changes how we produce character formatting keywords.

1. Character formatting is now collected separately from block-level;
2. It is stored in an SfxItemSet, and newer properties overwrite older
   propertiew with the same which id, avoiding some redundancy;
3. The generation of the actual RTF markup is therefore delayed, which
   means that the "current" position in the exported document (as seen
   e.g. by MSWordExportBase::HasItem) could change by that time. Since
   handling some properties needs other accompanying properties  (e.g.,
   handling RES_CHRATR_UNDERLINE needs RES_CHRATR_WORDLINEMODE), those
   accompanying items are collected together with the primary item.
4. When collecting the accompanying data, side effects could appear if
   they are put into the same SfxItemSet.  E.g., RES_CHRATR_ESCAPEMENT
   needs RES_CHRATR_FONTSIZE;  but the latter, when present in the set,
   will itself create RTF markup for font size. So it must be added in
   a way to avoid output of its RTF markup, when the run itself has no
   explicit font size set. For that, a parent set is introduced; items
   stored there will be reachable, when "search in parent" is used.

The function creating the resulting RTF markup tries to follow specs wrt
the order and use of associated keywords for scripts. Of course, I could
be wrong in the interpretation of the spec. At least some documents that
I tested using Word seem to improve with this change.

Change-Id: I3761c4c84f137be8ff9de0eec5ed8b229146325a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188427
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>