tdf#165564: refactor the "Removes additional spaces in export" part
Commit 90ce84a5dd371c804af19e97896a5e19a3ebf8ca (tdf#165564 RTF:Fix
export and import of footnotes/endnotes, 2025-05-08) workarounded
an extra space added after \super keyword, when EndRunProperties()
didn't produce an output. To do that, it introduced a return value
from EndRunProperties, checked in TextFootnote_Impl.

This change simplifies this, reverting a part of the said commit,
which added that return value, in favor of dropping an unnecessary
space after {\super, which was retained for some reason in commit
49877dc91b2f91baa656facd462ac1b1e832f182 (sw:rtf - timely export
of footnote char properties, 2017-08-24). There is no need in the
space between keywords - only between the last keyword and text.

Change-Id: If13bf760feb9cb42cb39fc7de39eed6944a869ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188433
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>