tdf#44729 sdext,pdfimport: Hide transparency groups 'for soft mask'
In PDF, 'Transparency groups' are used to group objects for compositing.
One variant of those is a transparency group that doesn't actually
get directly composited, but is used as a 'soft mask' to modulate
the effects of another group when the other group is composited.

So far we've not done anything for transparency groups, where
softmasks are used that means the softmask gets rendered as a visible
object - often a big semi-transparent polygon covering the rest.

Add 'group' elements into the tree for Transparency Groups and mark
them.
During tree walks skip the 'for soft mask' transparency groups.

This improves tdf#44729 which prior to this patch is mostly
seen as a big black page because it's covered with the soft mask
misrendering.

Since it also stored the transparency group it also means we can
attack the real rendering of those later.

Change-Id: Ieb1eb56b15d8def5c842c1d84349173ccca630dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188455
Reviewed-by: David Gilbert <freedesktop@treblig.org>
Tested-by: Jenkins