Related tdf#97694 Don't require signed document in Base
Macros were disabled in Base when calls to macros were detected in
the Base document - since the document is not signed.

However, Base does not support document signatures at this stage.

So relax that requirement to require a signed document when detecting
calls to signed macros from an unsigned document - at least for Base.

Change-Id: I129f3ffb2607a1c5eebf4795fcd1638c206aea22
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188458
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins