Related: tdf#167222 sw floattable: fix split of fly and its heading text w/ ftn
Open the updated bugdoc, which now has a footnote as well: the table
goes to page 2 in Word, but to page 3 in Writer.

This time what seems to happen is that the footnote container results in
less body frame space, so the entire floating table + anchor text moves
to page 3 and never moves back, even after the footnote moved to page 3.

Fix this by extending SwContentFrame::MakeAll() to be consistent with
the SwTextFrameBreak ctor, so it also ignores "don't break" / "keep with
next" when that would potentially result in not allowing a page break
between the floating table and its anchor text.

Change-Id: I3161fb6896909b32996fe83acd8d6772f7b66af6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188469
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188694