tdf#130857 qt weld: No longer use separate col for "expander toggle"
Change the approach to handling the special "expander toggle"
that is documented in the weld::TreeView API:

    // col index -1 sets the expander toggle, enable_toggle_buttons must have been called to create that column
    virtual void set_toggle(const TreeIter& rIter, TriState bOn, int col = -1) = 0;
    // col index -1 gets the expander toggle, enable_toggle_buttons must have been called to create that column
    virtual TriState get_toggle(const TreeIter& rIter, int col = -1) const = 0;

Instead of using a separate column for it, instead
always set the checked state for the existing column
at index 0.

This is similar to the approach for the "expander image"
implemented in

    commit a838839e4d191c62a9396ee65af9a116a617160f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jul 25 07:59:01 2025 +0200

        tdf#130857 qt weld: Support setting TreeView "expander image"

and (similar to what is described in the above-mentioned commit
for the expander image) not using a separate column will allow
having a column title for the "expander toggle" and the first
"normal" column together, which is e.g. relevant for the
document recovery dialog (UI file:
svx/uiconfig/ui/docrecoveryrecoverdialog.ui), where both,
the checkbox and the document name are supposed to be
in a column titled "Recover Document" together.

Support for that dialog will be declared in an upcoming commit.

In line with the above-mentioned commit, using the
column at index 0 implies that it's not possible to
set another check state for the same column. Therefore,
add a corresponding assert to detect any cases where
QtInstanceTreeView::get_toggle or
QtInstanceTreeView::set_toggle are called with index 0
if the special expander toggle is enabled.
If any such cases exist, they will have to be addressed
once they show up. (Depending on the use case, reworking
the implementation of the specific dialog could be a
potential option, e.g. introducing a separate column.)

Change-Id: I65b1c94aec27985cbcc3a173a5f8bcb50e6aa167
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188533
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>