tdf#130857 qt weld: Forward DrawingArea context menu event
Intercept QContextMenuEvent for the QtInstanceDrawingArea
widget, convert to a corresponding vcl CommandEvent
and and call the corresponding handler.

This will e.g. be needed to make the context menu in the
Base "Tools" -> "SQL" dialog work when right-clicking into
the "Command to execute" edit view.

In a WIP branch where support for that dialog is
declared, this now results in - the not yet implemented -
QtInstanceMenu::popup_at_rect getting called when
on right click.

Change-Id: If8cfff809878d43e6b45b0b8317d1bb8738397d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188540
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>