tdf#130857 qt weld: Support edit menu
This means that a native Qt menu is now used for that menu
when using the qt5 or qt6 VCL plugin if the menu has a native
Qt widget parent and LO was started with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Together with previous commit

    Change-Id: I66d273895e55698427343caf688b88b2e56e3993
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 29 20:10:37 2025 +0200

        tdf#130857 qt weld: Implement QtInstanceMenu::popup_at_rect

and various other changes on a local WIP branch, this
makes the context menu in the Base "Tools" -> "SQL"
dialog show when right-clicking into the "Command to execute"
edit view.

Change-Id: Idf2a212ea27929342f2f965d25a791c53b20f3ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188542
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>