tdf#130857 cui a11y: Set mnemnic widget in "Arrow Styles" tab page
This e.g. makes the Orca screen reader announce
the "Arrow style" of the corresponding combobox
when it receives focus when using the gtk3 VCL plugin.

To reproduce the dialog:

* start Orca
* start Draw
* insert a shape (e.g. a rectangle)
* right-click on newly inserted shape to open context menu
* select "Line" context menu entry
* go to "Arrow Styles" tab
* press tab to move focus to the "Arrow style" combobox

In addition, it also makes the label show the proper
text in a WIP branch where support for using native
widgets is declared for that dialog with the qt6 VCL
plugin, as otherwise a literal "Arrow &style" is
shown when an accelerator is set without a mnemonic
widget.

Change-Id: I85e19079d4d3d07861c581f6ff4123a96014b2f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188554
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins