tdf#130857 qt weld: Improve QtInstanceWidget::{is,get}_visible
QWidget::isVisible only returns true when the widget
is actually visible on screen, i.e. itself and all
widgets up to the top-level are marked visible, which matches
the weld::Widget::is_visible/gtk_widget_is_visible [1]
semantics, so use it to implement that.

For QtInstanceWidget::get_visible/gtk_widget_get_visible [2],
the question is whether only the widget itself is marked
visible, i.e. it would be shown on screen when all parents
up to the toplevel were marked visible as well.

Reimplement that to return whether all are already visible
or QWidget::isVisibleTo returns true when called with
the parent as a param. Quoting the doc [3]:

> Returns true if this widget would become visible if ancestor is shown;
> otherwise returns false.
>
> The true case occurs if neither the widget itself nor any parent up to
> but excluding ancestor has been explicitly hidden.

Without this commit in place, the "Settings" frame
in the chart "Data Series" dialog would not be shown
in a WIP branch declaring support for that dialog
due to the following logic as the `get_visible()`
implementations for `m_xMTGap` and `m_xMTOverlap`
returned false as long as the dialog wasn't shown
in SchOptionTabPage::AdaptControlPositionsAndVisibility:

    void SchOptionTabPage::AdaptControlPositionsAndVisibility()
    {
        m_xGrpAxis->set_visible(m_bProvidesSecondaryYAxis);
        m_xGrpBar->set_visible(m_bProvidesOverlapAndGapWidth);
        m_xCBConnect->set_visible(m_bProvidesBarConnectors);

        if (!m_xMTGap->get_visible() && !m_xMTOverlap->get_visible())
            m_xGrpBar->hide();
    }

To trigger the dialog:

* start Writer
* press the "Insert Chart" toolbar button to insert a chart
* double-click on one of the bars in the chart

[1] https://docs.gtk.org/gtk3/method.Widget.is_visible.html
[2] https://docs.gtk.org/gtk3/method.Widget.get_visible.html

Change-Id: I03584fe96b62bb0d58bd27f7a4f49e4c0266ca1a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188596
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins