Related: tdf#167222 sw floattable: fix split of fly and heading text fit check
Open the bugdoc, the only floating table is at the top of page 4 in the
document, while it's at the bottom of page 3 in Word.

This time the layout decides in SwContentFrame::WouldFit_() that the
floating table and its anchor text can't fit page 3 and instead of
trying to split, it moves both to page 4.

Fix the problem similar to what happens already in
SwContentFrame::MakeAll() and SwTextFrameBreak, by ignoring "don't
break" / "keep with next" when that would potentially result in not
allowing a page break between the floating table and its anchor text.

Also extract this "ignore value for split fly anchors" logic to a new
SwContentFrame::IgnoringSplitFlyAnchor() to avoid duplication.

Change-Id: Ic31f0e34c75d978b4d8b6cbe2451ee7e15ef2f4d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188604
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188695