tdf#130857 xmlsecurity: Fix path for unvalidated cert icon
The path was missing an "s" in "xmlsecurity".

As a consequence, the icon couldn't be loaded, and thus
couldn't be displayed.

For the qt6 VCL plugin with SAL_VCL_QT_USE_WELDED_WIDGETS=1,
it triggered first a warning, and then an assert:

    warn:vcl:3089384:3089384:vcl/source/bitmap/BitmapEx.cxx:103: BitmapEx::BitmapEx(): could not load image xmlecurity/res/notcertificate_40x56.png via icon theme breeze
    soffice.bin: /home/michi/development/git/libreoffice/include/vcl/qt/QtUtils.hxx:48: QPixmap toQPixmap(const BitmapEx &): Assertion `!aPixmap.isNull() && "Failed to create icon pixmap"' failed.

Support for that dialog was added in previous commit

    Change-Id: Ib0ec32ff50744f4e265a743f2b0e5f240253ee5f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jul 30 11:51:12 2025 +0200

        tdf#130857 qt weld: Support "View Certificate" dialog

The issue could be triggered as follows when a GPG certificate
that is not considered validated is used:

* start Writer
* save doc
* "File" -> "Digital Signatures" -> "Digital Signatures"
* press "Sign Document"
* select an existing certificate/GPG key
* press "View Certificate"

Now, the icon is shown as expected with all of gtk3, gen and qt6.

Path was spelled the wrong way way ever since

    commit 7bd87eb30ca5df683fa3bd94ca5be9dbca362c03
    Date:   Wed Apr 26 09:35:24 2017 +0100

        remove xmlsecurity bitmaps from .src files

Change-Id: Icbe2f56dc0b831a0d56fc2707fed6480cb0b9bae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188600
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>