tdf#123159 don't show 'open hyperlink' in context menu unneccessarily
14396a5361ef72afc99033d4cac29aaeb1ffa2d8 extended
.uno:OpenHyperlinkOnCursor to open all the hyperlinks in the current
cell if not in edit mode (to make hyperlinks keyboard accessible), but
i forgot to add `SID_OPEN_HYPERLINK` entry to `ScDocShell::GetState()`
which caused this bug.

now we check for both the cases if the cell has hyperlink(s) or not
and disable the slot if it doesn't so that we don't re-run the logic
in the execute function unneccessary.

Change-Id: I378996d776903ab53d6370d19ace9d3c730f3e85
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188620
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@collabora.com>
(cherry picked from commit fe5518051d1446627818c6e9704e404f04d3eb42)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188736
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>