tdf#167647 add tabs to find and replace dialog
add two toggle buttons on top of the "find replace dialog" and make
them behave as tabs i.e. when replace is clicked, show the replace
controls.

when the dialog first shows up, we hide the replace controls and when
the user switches to replace tab, we show the replace controls, find
controls are always visible.

i thought through various other ways it could have been done (most of
them being more complicated than the current solution).

1. "could we create one tab and just return a shared pointer to it
for different tabs so that they share the page". this requires us to
return a shared pointer from TabPageClass::Create which is passed in
the AddTabPage function. i thought "could we internally store a raw
pointer and then in the Create function of the tab page, we return
the unique_ptr to it." this doesn't seem like a good idea as when the
first tab page is destructed, the raw pointer in the Create function
and the unique pointers in other tab pages will become invalid leading
to double free.

2. "what if i add the frame only to the first tab and in the code
set it as a child of the second tab as well, using something like
set_child" -> weld doesn't have any such function as far as i could
search.

3. convert the current find-replace dialog into a tab page and add two
different instances of it to the dialog. then transfer the data to
the other tab on tab switch. this would require addition of a lot of
getters and setters and might introduce bugs.

Change-Id: I49c5640f39262578e31caee839c2059bb86610a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188644
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>