tdf#167194 sw redline reinstate: fix handling of self-inserts
Create an insert in a document, then Edit -> Track changes -> Reinstate,
you get an empty document instead of an "insert-then-delete" as
expected.

This happens because
sw::DocumentRedlineManager::PreAppendDeleteRedline() has a call to
IsOwnRedline(), which compresses the insert and the delete, which is
wanted in general for self-inserts, but not while reinstating.

Fix the problem by disabling redline compression in
SwEditShell::ReinstatePaM() for the insert case.

This requires extending
sw::DocumentContentOperationsManager::DeleteAndJoinWithRedlineImpl()
which probably just wants to turn on redline recording, but dropped
RedlineFlags::DontCombineRedlines while doing so.

Change-Id: I629d4d65dc966af934c7ecc672925d0fc11a3ef9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188734
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>