tdf#130857 weld: No longer let ScrolledWindow subclass Container
While for GTK 3, GtkScrolledwindow subclasses GtkContainer
(via GtkBin), the Qt implementation of weld::Container,
QtInstanceContainer, depends on a widget that has a
layout set.
The whole concept doesn't fit too well for the QScrollArea
widget that QtInstanceScrolledWindow uses, which is based
on the idea of providing scrolling capabilities for a single
child widget.

As far as I can see, there is no particular reason for
weld::ScrolledWindow to subclass weld::Container and
none of the ScrolledWindow instances used in the code
seems to use the weld::Container API, so drop that
inheritance for weld::ScrolledWindow.

The GTK and VCL implementations remain unchanged.

Change-Id: I7ce5d6296b9c40cee8ec8eb6dcfb9b1fb81d4ca3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188759
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>