tdf#130857 weld: Rename signals for ScrolledWindow scroll value change
A GtkAdjustment has multiple properties, and
the value (which corresponds to the current
position of the slider in the horizontal/vertical
scrollbar for weld::ScrolledWindow) is only one of them.

In fact, there is a GtkAdjustment::changed signal [1] in
GTK that is emitted whenever anything *except* for the value
changes, while GtkAdjustment::value_changed [2] is emitted
when the value changes.

For weld::ScrolledWindow, the semantics is the latter
(s. e.g. how the GTK implementation emits the signal
when it receives the "value-changed" signal for the
corresponding GtkAdjustment).

Adjust the naming accordingly, to make clear that this
is only about value changes.

[1] https://docs.gtk.org/gtk3/method.Adjustment.changed.html
[2] https://docs.gtk.org/gtk3/method.Adjustment.value_changed.html

Change-Id: Ie6ce7ff7f69725ddbf7ac4cb2e60c391a02ef1f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188761
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>